# pkg/docs/swaggers/zpotify_auth_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "zpotify_auth_service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "AuthAPI"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/auth/async": {
      "post": {
        "operationId": "AuthAPI_AuthAsync",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/zpotify_apiAuthViaAsyncResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of zpotify_apiAuthViaAsyncResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/zpotify_apiAuthViaAsyncRequest"
            }
          }
        ],
        "tags": [
          "AuthAPI"
        ]
      }
    },
    "/api/auth/auth_methods": {
      "get": {
        "operationId": "AuthAPI_GetAuthMethods",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/zpotify_apiGetAuthMethodsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "AuthAPI"
        ]
      }
    },
    "/api/auth/refresh_token": {
      "post": {
        "operationId": "AuthAPI_RefreshToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/zpotify_apiRefreshResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/zpotify_apiRefreshRequest"
            }
          }
        ],
        "tags": [
          "AuthAPI"
        ]
      }
    },
    "/api/auth/sync": {
      "post": {
        "operationId": "AuthAPI_Auth",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/zpotify_apiAuthResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/zpotify_apiAuthRequest"
            }
          }
        ],
        "tags": [
          "AuthAPI"
        ]
      }
    }
  },
  "definitions": {
    "AuthLogPass": {
      "type": "object",
      "properties": {
        "login": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "zpotify_apiAuthData": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "accessExpiresAt": {
          "type": "string",
          "format": "date-time"
        },
        "refreshExpiresAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "zpotify_apiAuthMethods": {
      "type": "string",
      "enum": [
        "AuthMethod_NotSpecified",
        "AuthMethod_Offline",
        "AuthMethod_Telegram"
      ],
      "default": "AuthMethod_NotSpecified"
    },
    "zpotify_apiAuthRequest": {
      "type": "object",
      "properties": {
        "logPass": {
          "$ref": "#/definitions/AuthLogPass"
        }
      }
    },
    "zpotify_apiAuthResponse": {
      "type": "object",
      "properties": {
        "authData": {
          "$ref": "#/definitions/zpotify_apiAuthData"
        }
      }
    },
    "zpotify_apiAuthViaAsyncRequest": {
      "type": "object",
      "title": "By default considering auth goes via telegram"
    },
    "zpotify_apiAuthViaAsyncResponse": {
      "type": "object",
      "properties": {
        "authUuid": {
          "type": "string"
        },
        "authData": {
          "$ref": "#/definitions/zpotify_apiAuthData"
        }
      }
    },
    "zpotify_apiGetAuthMethodsResponse": {
      "type": "object",
      "properties": {
        "authMethods": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/zpotify_apiAuthMethods"
          }
        }
      }
    },
    "zpotify_apiRefreshRequest": {
      "type": "object",
      "properties": {
        "refreshToken": {
          "type": "string"
        }
      }
    },
    "zpotify_apiRefreshResponse": {
      "type": "object",
      "properties": {
        "authData": {
          "$ref": "#/definitions/zpotify_apiAuthData"
        }
      }
    }
  }
}